Added support for secondary norminette command
Added a command2 and fileregex2 option to be able to run a secondary command, like norminette+.
Modified documentation to reflect the new feature.
Module now explicitly checks for each command if it is non-empty before it runs the command.
Code could be improved since there is a bunch of reused code, but it's functional...